Fixes circular imports (badly)
