+++ fia_api/web/api/teacher/views.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter, Depends
 
 from fia_api.db.models.user_model import UserModel
-from fia_api.web.api.teacher.schema import TeacherConverseRequest
+from fia_api.web.api.teacher.schema import ConversationResponse, TeacherConverseRequest
 from fia_api.web.api.teacher.utils import get_response, initialize_conversation
-from fia_api.web.api.user.schema import AuthenticatedUser, ConversationResponse
+from fia_api.web.api.user.schema import AuthenticatedUser
 from fia_api.web.api.user.utils import get_current_user
 
 router = APIRouter()
